Fix implicit-function parameter scoping using global scope
current_scope was used when creating a type_ref_func. This was being
done at fold-time too (for implicit function decls). This meant an
implicit function incorrectly could end up with global scope for
its argument-symbol-table.

So when we came to set are_params for function type_ref:s, we would
mark the global scope as a parameter scope, which meant any decls
declared therein would end up with sym_arg as their symbol type.

This then meant that when we generated calls to functions, we would
try to access an argument, rather than a global symbol (sym_arg
instead of sym_global / sym_local-with-function).