Fix rewrite of pure inline functions to static/extern
We now allow (as a special case) a extern or static declaration of
an inline function to alter the original's storage, so that

inline int f() { ... }
extern inline int f();

will cause a global symbol for "f" (and code) to be emitted.

Also corrected warning for emission of pure inline functions - we
now warn at the symtable checking stage, after we've allowed all
late-declarations to affect the inline functions definitions.